test(e2e): expand harness-perm — absorb 2 standalone probes + fix selector regression
- Absorb probe-e2e-permission-auto-and-titles into harness as
  permission-auto-and-titles (store + i18n + IPC contract for 'auto' mode +
  per-tool prompt titles).
- Absorb probe-e2e-permission-reject-stops-agent into harness as
  permission-reject-stops-agent (deny IPC fires once, no retry, denial
  trace remains in chat).
- Fix latent regression: harness used `text=Permission required` heading
  selector, broken since #288 swapped the heading to per-tool titles
  ("Allow this bash command?"). Switched all 12 cases to the stable
  `[role="alertdialog"]` container selector. Pre-fix: 9/10 cases failed
  on origin/working baseline. Post-fix: 12/12 pass.
- Delete the 2 absorbed standalone probe files (no breadcrumb files; #72
  no-skipped-e2e rule).

Reduces probe-e2e file count 65 -> 63; saves ~2 electron launches per
run-all-e2e (~30s wall).

Reverse-verify (per-case sanity):
- permission-auto-and-titles: changed `modeAutoLabel: 'Auto'` to 'BROKEN'
  in src/i18n/locales/en.ts -> case FAILED ("autoLabel expected Auto, got
  BROKEN"). Reverted -> PASSES.
- permission-reject-stops-agent: removed `onReject?.()` call from the 'n'
  hotkey handler in src/components/PermissionPromptBlock.tsx -> case
  FAILED ("block still visible after pressing N"). Reverted -> PASSES.
