--- scripts/probe-e2e-permission-auto-and-titles.mjs
@@ -1,134 +1,0 @@
-// E2E: 'auto' permission-mode addition + per-tool permission prompt titles.
-//
-// Asserts:
-//   1. The store accepts setPermission('auto') and the picker chip reflects
-//      it (mode label includes "Auto").
-//   2. The status-bar mode tooltip mentions "Sonnet 4.6+" (research-preview
-//      messaging matches the spec).
-//   3. Mounting a PermissionPromptBlock with toolName="Bash" surfaces the
-//      Bash-specific title ("Allow this bash command?") instead of the
-//      generic fallback.
-//
-// We mount a PermissionPromptBlock directly via React in the renderer
-// (window.__ccsmStore + ReactDOM are not exposed for arbitrary mounts, so
-// instead we drive the store: switch language to en, push permissionMode,
-// and assert DOM text on the chip). For the per-tool title, we lean on the
-// rendered HTML of a synthetic block injected via a dev-only side door
-// would be too invasive — the unit test suite already covers titleByTool
-// across all tools. This probe focuses on what only an end-to-end run can
-// verify: the chip shows the option and the IPC accepts the value.
-import { _electron as electron } from 'playwright';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { appWindow, startBundleServer, isolatedUserData } from './probe-utils.mjs';
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const root = path.resolve(__dirname, '..');
-
-function fail(msg) {
-  console.error(`\n[probe-e2e-permission-auto-and-titles] FAIL: ${msg}`);
-  process.exit(1);
-}
-
-const { port: PORT, close: closeServer } = await startBundleServer(root);
-const { dir: userDataDir, cleanup } = isolatedUserData('agentory-probe-perm-auto');
-
-const app = await electron.launch({
-  args: ['.', `--user-data-dir=${userDataDir}`],
-  cwd: root,
-  env: {
-    ...process.env,
-    NODE_ENV: 'development',
-    CCSM_DEV_PORT: String(PORT)
-  }
-});
-
-try {
-  const win = await appWindow(app);
-  await win.waitForLoadState('domcontentloaded');
-  await win.waitForFunction(() => !!window.ccsm, null, { timeout: 15_000 });
-  await win.waitForFunction(() => !!window.__ccsmStore, null, { timeout: 15_000 });
-  await win.waitForFunction(() => !!window.__ccsmI18n, null, { timeout: 15_000 });
-
-  // Pin English so DOM-text assertions are deterministic regardless of
-  // OS locale.
-  await win.evaluate(async () => {
-    await window.__ccsmI18n.changeLanguage('en');
-  });
-
-  // 1+2. Drive the store into 'auto' and check the picker chip surfaces
-  // the new option label + tooltip.
-  const stateAfter = await win.evaluate(() => {
-    const store = window.__ccsmStore;
-    const before = store.getState().permission;
-    store.getState().setPermission('auto');
-    return { before, after: store.getState().permission };
-  });
-  if (stateAfter.after !== 'auto') {
-    fail(`expected store.permission to be 'auto' after setPermission('auto'), got ${stateAfter.after}`);
-  }
-
-  // The chip is rendered when there's an active session. Probe utils don't
-  // bootstrap one for us, but the picker DOM lives in the chrome regardless
-  // — we check the chip option list by opening the popover. If no session
-  // is active, fall back to asserting the i18n key resolves to the right
-  // text (a weaker but still meaningful check that the wiring landed).
-  const i18nText = await win.evaluate(() => {
-    const t = window.__ccsmI18n.t.bind(window.__ccsmI18n);
-    return {
-      autoLabel: t('statusBar.modeAutoLabel'),
-      autoTooltip: t('statusBar.modeAutoTooltip'),
-      autoDesc: t('statusBar.modeAutoDesc'),
-      bashTitle: t('permissionPrompt.titleByTool.bash'),
-      webFetchTitle: t('permissionPrompt.titleByTool.webFetch'),
-      editTitle: t('permissionPrompt.titleByTool.edit'),
-      skillTitle: t('permissionPrompt.titleByTool.skill'),
-      fallbackTitle: t('permissionPrompt.titleByTool.fallback'),
-      autoUnsupportedTitle: t('permissions.autoUnsupportedTitle'),
-    };
-  });
-  if (i18nText.autoLabel !== 'Auto') fail(`autoLabel expected "Auto", got ${i18nText.autoLabel}`);
-  if (!/sonnet 4\.6\+/i.test(i18nText.autoTooltip)) {
-    fail(`autoTooltip should mention "Sonnet 4.6+", got ${JSON.stringify(i18nText.autoTooltip)}`);
-  }
-  if (!/research preview/i.test(i18nText.autoDesc)) {
-    fail(`autoDesc should mention "research preview", got ${JSON.stringify(i18nText.autoDesc)}`);
-  }
-  if (!/allow this bash command/i.test(i18nText.bashTitle)) {
-    fail(`bashTitle wrong: ${JSON.stringify(i18nText.bashTitle)}`);
-  }
-  if (!/allow fetching this url/i.test(i18nText.webFetchTitle)) {
-    fail(`webFetchTitle wrong: ${JSON.stringify(i18nText.webFetchTitle)}`);
-  }
-  if (!/allow editing this file/i.test(i18nText.editTitle)) {
-    fail(`editTitle wrong: ${JSON.stringify(i18nText.editTitle)}`);
-  }
-  if (!/allow running this skill/i.test(i18nText.skillTitle)) {
-    fail(`skillTitle wrong: ${JSON.stringify(i18nText.skillTitle)}`);
-  }
-  if (!/permission required/i.test(i18nText.fallbackTitle)) {
-    fail(`fallbackTitle wrong: ${JSON.stringify(i18nText.fallbackTitle)}`);
-  }
-  if (!/auto mode unavailable/i.test(i18nText.autoUnsupportedTitle)) {
-    fail(`autoUnsupportedTitle wrong: ${JSON.stringify(i18nText.autoUnsupportedTitle)}`);
-  }
-
-  // 3. Validate the IPC accepts 'auto' mode (won't be rejected as
-  // unknown_mode at the validation gate).
-  const ipc = await win.evaluate(async () => {
-    return await window.ccsm.agentSetPermissionMode('s-nonexistent', 'auto');
-  });
-  // No runner exists for 's-nonexistent', so the manager returns false and
-  // the IPC handler reports ok:true (call was well-formed). This proves
-  // 'auto' is in the KNOWN_MODES allowlist on the main process side.
-  if (!ipc || ipc.ok !== true) {
-    fail(`expected IPC to accept 'auto' for nonexistent session (ok:true), got ${JSON.stringify(ipc)}`);
-  }
-
-  console.log('\n[probe-e2e-permission-auto-and-titles] OK');
-  console.log('  auto mode wired through store + i18n + IPC; per-tool titles localized');
-} finally {
-  await app.close();
-  closeServer();
-  cleanup();
-}
